Поправил имена и разметку.
Сереализуется без ошибок в сравнении с файлом из примера.

+++ TopoR/Commons.cpp
@@ -1,6 +1,6 @@
 #include "Commons.h"
 namespace TopoR {
-void Coord::Shift(float x_, float y_) {
+void Coord::Shift(double x_, double y_) {
     x += x_;
     y += y_;
 }
@@ -8,13 +8,13 @@ void Coord::UnitsConvert(dist in_units, dist out_units) {
     x = Ut::UnitsConvert(x, in_units, out_units);
     y = Ut::UnitsConvert(y, in_units, out_units);
 }
-void SegmentLine::Shift(float x, float y) {
+void SegmentLine::Shift(double x, double y) {
     End.Shift(x, y);
 }
 void SegmentLine::UnitsConvert(dist in_units, dist out_units) {
     End.UnitsConvert(in_units, out_units);
 }
-void SegmentArcCCW::Shift(float x, float y) {
+void SegmentArcCCW::Shift(double x, double y) {
     Center.Shift(x, y);
     End.Shift(x, y);
 }
@@ -22,7 +22,7 @@ void SegmentArcCCW::UnitsConvert(dist in_units, dist out_units) {
     Center.UnitsConvert(in_units, out_units);
     End.UnitsConvert(in_units, out_units);
 }
-void SegmentArcByMiddle::Shift(float x, float y) {
+void SegmentArcByMiddle::Shift(double x, double y) {
     Middle.Shift(x, y);
     End.Shift(x, y);
 }
@@ -30,7 +30,7 @@ void SegmentArcByMiddle::UnitsConvert(dist in_units, dist out_units) {
     Middle.UnitsConvert(in_units, out_units);
     End.UnitsConvert(in_units, out_units);
 }
-void ArcCCW::Shift(float x, float y) {
+void ArcCCW::Shift(double x, double y) {
     Start.Shift(x, y);
     Center.Shift(x, y);
     End.Shift(x, y);
@@ -40,7 +40,7 @@ void ArcCCW::UnitsConvert(dist in_units, dist out_units) {
     Center.UnitsConvert(in_units, out_units);
     End.UnitsConvert(in_units, out_units);
 }
-void ArcByAngle::Shift(float x, float y) {
+void ArcByAngle::Shift(double x, double y) {
     Start.Shift(x, y);
     End.Shift(x, y);
 }
@@ -48,7 +48,7 @@ void ArcByAngle::UnitsConvert(dist in_units, dist out_units) {
     Start.UnitsConvert(in_units, out_units);
     End.UnitsConvert(in_units, out_units);
 }
-void ArcByMiddle::Shift(float x, float y) {
+void ArcByMiddle::Shift(double x, double y) {
     Start.Shift(x, y);
     Middle.Shift(x, y);
     End.Shift(x, y);
@@ -58,7 +58,7 @@ void ArcByMiddle::UnitsConvert(dist in_units, dist out_units) {
     Middle.UnitsConvert(in_units, out_units);
     End.UnitsConvert(in_units, out_units);
 }
-void Circle::Shift(float x, float y) {
+void Circle::Shift(double x, double y) {
     Center.Shift(x, y);
 }
 void Circle::UnitsConvert(dist in_units, dist out_units) {
@@ -68,7 +68,7 @@ void Circle::UnitsConvert(dist in_units, dist out_units) {
 bool Line::ShouldSerialize_Dots() {
     return Dots.size();
 }
-void Line::Shift(float x, float y) {
+void Line::Shift(double x, double y) {
     for(int i{}; i < Dots.size(); i++)
         Dots[i].Shift(x, y);
 }
@@ -79,7 +79,7 @@ void Line::UnitsConvert(dist in_units, dist out_units) {
 bool Polyline::ShouldSerialize_Segments() {
     return Segments.size();
 }
-void Polyline::Shift(float x, float y) {
+void Polyline::Shift(double x, double y) {
     Start.Shift(x, y);
     // for(int i{}; i < Segments.size(); i++)
     //     (std::dynamic_pointer_cast<IBaseSegment>(Segments[])).Shift(x, y);
@@ -89,7 +89,7 @@ void Polyline::UnitsConvert(dist in_units, dist out_units) {
     // for(int i{}; i < Segments.size(); i++)
     //     (std::dynamic_pointer_cast<IBaseSegment>(Segments[])).UnitsConvert(in_units, out_units);
 }
-void TrackArcCW::Shift(float x, float y) {
+void TrackArcCW::Shift(double x, double y) {
     Center.Shift(x, y);
     End.Shift(x, y);
 }
@@ -97,7 +97,7 @@ void TrackArcCW::UnitsConvert(dist in_units, dist out_units) {
     Center.UnitsConvert(in_units, out_units);
     End.UnitsConvert(in_units, out_units);
 }
-void TrackLine::Shift(float x, float y) {
+void TrackLine::Shift(double x, double y) {
     End.Shift(x, y);
 }
 void TrackLine::UnitsConvert(dist in_units, dist out_units) {
@@ -106,7 +106,7 @@ void TrackLine::UnitsConvert(dist in_units, dist out_units) {
 void Thermal::UnitsConvert(dist in_units, dist out_units) {
     spokeWidth = Ut::UnitsConvert(spokeWidth, in_units, out_units);
 }
-void Detail::Shift(float x, float y) {
+void Detail::Shift(double x, double y) {
     // if((std::dynamic_pointer_cast<IBaseFigure>(Figure)) != nullptr)
     //     (std::dynamic_pointer_cast<IBaseFigure>(Figur)).Shift(x, y);
 }
@@ -118,13 +118,13 @@ void Detail::UnitsConvert(dist in_units, dist out_units) {
 bool Text::getMirrorSpecified() const {
     return mirror != Bool::off;
 }
-void Text::Shift(float x, float y) {
+void Text::Shift(double x, double y) {
     Org.Shift(x, y);
 }
 void Text::UnitsConvert(dist in_units, dist out_units) {
     Org.UnitsConvert(in_units, out_units);
 }
-float Ut::UnitsConvert(float value, dist in_units, dist out_units) {
+double Ut::UnitsConvert(double value, dist in_units, dist out_units) {
     double k;
     switch(in_units) {
     case dist::mkm: k = 0.001; break;
@@ -137,13 +137,13 @@ float Ut::UnitsConvert(float value, dist in_units, dist out_units) {
     default: k = 1; break;
     }
     switch(out_units) {
-    case dist::mkm: return static_cast<float>(value * k * 1000);
-    case dist::cm: return static_cast<float>(value * k * 0.1);
-    case dist::dm: return static_cast<float>(value * k * 0.01);
-    case dist::m: return static_cast<float>(value * k * 0.001);
-    case dist::mil: return static_cast<float>(value * k * 39.37007874015748);
-    case dist::inch: return static_cast<float>(value * k * 0.03937007874015748);
-    case dist::mm: return static_cast<float>(value * k);
+    case dist::mkm: return static_cast<double>(value * k * 1000);
+    case dist::cm: return static_cast<double>(value * k * 0.1);
+    case dist::dm: return static_cast<double>(value * k * 0.01);
+    case dist::m: return static_cast<double>(value * k * 0.001);
+    case dist::mil: return static_cast<double>(value * k * 39.37007874015748);
+    case dist::inch: return static_cast<double>(value * k * 0.03937007874015748);
+    case dist::mm: return static_cast<double>(value * k);
     default: return value;
     }
 }
