+++ TopoR/Layers.cpp
@@ -7,7 +7,7 @@ bool Layers::Layer::getThicknessSpecified() const {
     return {}; //    return type != layertype::Assy;
 }
 Layers::Layer::Layer() { }
-Layers::Layer::Layer(const std::string& name, layertype type, Bool compsOutline, float thickness) {
+Layers::Layer::Layer(const std::string& name, layertype type, Bool compsOutline, double thickness) {
     //    name = name;
     //    type = type;
     //    compsOutline = compsOutline;
@@ -23,14 +23,14 @@ bool Layers::ShouldSerialize_UnStackLayers() {
     return {}; //    return UnStackLayers.size();
 }
 bool Layers::LayerStackUpContains(LayerRef lref) {
-    return {}; //    return (StackUpLayers.empty() ? nullptr : StackUpLayers.Where([&](std::variant<> r) {
+    return {}; //    return (StackUpLayers.empty() ? nullptr : StackUpLayers.Where([&](std::variant<XML::Null, > r) {
     //                                                                 return r->name == lref->ReferenceName;
     //                                                             })
     //                                                   ->Count())
     //        > 0;
 }
 bool Layers::LayerUnStackContain(LayerRef lref) {
-    return {}; //    return (UnStackLayers.empty() ? nullptr : UnStackLayers.Where([&](std::variant<> r) {
+    return {}; //    return (UnStackLayers.empty() ? nullptr : UnStackLayers.Where([&](std::variant<XML::Null, > r) {
     //                                                                 return r->name == lref->ReferenceName;
     //                                                             })
     //                                                   ->Count())
